// src/interfaces/constantsInterfaces.ts
export interface SkillsLogoJsonInterface {
  NAME: string
  CATEGORY: string
  COMPONENT: string
  GEOMETRY: {
    POSITION: {
      X: number
      Y: number
      Z: number
    }
    ROTATION: {
      X: number
      Y: number
      Z: number
    }
    SCALE: {
      X: number
      Y: number
      Z: number
    }
  }
}